add: 추천 굿생 게시글 뷰
