+++ feature/community-page/src/main/java/com/godlife/community_page/stimulus/StimulusPostScreen.kt
@@ -3,7 +3,6 @@ package com.godlife.community_page.stimulus
 import android.annotation.SuppressLint
 import android.graphics.Bitmap
 import android.graphics.drawable.Drawable
-import android.util.Log
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.expandVertically
@@ -26,7 +25,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.grid.GridCells
@@ -37,20 +35,17 @@ import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material.icons.filled.Warning
 import androidx.compose.material.icons.rounded.Create
 import androidx.compose.material.icons.rounded.Menu
 import androidx.compose.material.icons.rounded.Search
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
@@ -90,14 +85,15 @@ import com.bumptech.glide.request.target.CustomTarget
 import com.bumptech.glide.request.transition.Transition
 import com.godlife.community_page.BuildConfig
 import com.godlife.community_page.R
-import com.godlife.community_page.navigation.PostDetailRoute
 import com.godlife.community_page.navigation.StimulusPostDetailRoute
 import com.godlife.community_page.search.stimulus.StimulusSearchScreen
+import com.godlife.community_page.stimulus.recommended_post.RecommendedStimulusPostContent
+import com.godlife.community_page.stimulus.recommended_post.RecommendedStimulusPostContentPreview
 import com.godlife.create_post.stimulus.CreateStimulusPostScreen
 import com.godlife.designsystem.theme.GodLifeTheme
 import com.godlife.designsystem.theme.GrayWhite
 import com.godlife.designsystem.theme.OpaqueDark
-import com.godlife.network.model.StimulusPost
+import com.godlife.designsystem.theme.PurpleMain
 import com.godlife.network.model.StimulusPostList
 import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.delay
@@ -168,7 +164,7 @@ fun StimulusPostScreen(
                     ) {
 
                         item {
-                            StimulusPostContent1Preview(navController = navController)
+                            RecommendedStimulusPostContent(navController = navController)
                         }
 
 
@@ -390,6 +386,66 @@ fun LatestStimulusItem(
     }
 }
 
+@Composable
+fun StimulusCoverItem(
+    modifier: Modifier = Modifier,
+    item: StimulusPostList
+){
+    Box(
+        modifier = modifier
+            .padding(10.dp)
+            .size(width = 200.dp, height = 250.dp)
+            .shadow(10.dp),
+        contentAlignment = Alignment.Center
+    ){
+        val bitmap: MutableState<Bitmap?> = remember { mutableStateOf(null) }
+
+        Glide.with(LocalContext.current)
+            .asBitmap()
+            .load(BuildConfig.SERVER_IMAGE_DOMAIN + item.thumbnailUrl)
+            .error(R.drawable.category3)
+            .into(object : CustomTarget<Bitmap>() {
+                override fun onResourceReady(resource: Bitmap, transition: Transition<in Bitmap>?) {
+                    bitmap.value = resource
+                }
+
+                override fun onLoadCleared(placeholder: Drawable?) {}
+            })
+
+        bitmap.value?.asImageBitmap()?.let { fetchedBitmap ->
+            Image(
+                bitmap = fetchedBitmap,
+                contentDescription = null,
+                contentScale = ContentScale.FillWidth,
+                modifier = modifier
+                    .fillMaxWidth()
+            )   //bitmap이 없다면
+        } ?: Image(
+            painter = painterResource(id = R.drawable.category3),
+            contentDescription = null,
+            contentScale = ContentScale.FillWidth,
+            modifier = modifier
+                .fillMaxWidth()
+        )
+
+        Box(
+            modifier = modifier
+                .fillMaxSize()
+                .padding(30.dp)
+                .background(color = OpaqueDark)
+                .padding(start = 20.dp, end = 20.dp, top = 10.dp, bottom = 10.dp),
+            contentAlignment = Alignment.Center
+        ){
+
+            Text(text = item.title,
+                style = TextStyle(color = Color.White, fontSize = 18.sp, fontWeight = FontWeight.Bold)
+            )
+
+        }
+
+    }
+}
+
 @SuppressLint("UnrememberedMutableInteractionSource")
 @Composable
 fun CustomExpandableFAB(
@@ -489,185 +545,6 @@ fun CustomExpandableFAB(
 
 }
 
-
-@Preview
-@Composable
-fun StimulusPostContent1Preview(
-    modifier: Modifier = Modifier,
-    navController: NavController? = null
-){
-
-    val item = listOf(
-        StimulusPostItem(
-            title = "이것은 제목!",
-            writer = "치킨 러버",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "나도 제목!",
-            writer = "초코 러버",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제목 등장",
-            writer = "라면 좋아",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제모오오오옥",
-            writer = "헬로우",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        )
-    )
-
-    var width by remember {
-        mutableStateOf(0.dp)
-    }
-
-    val localDensity = LocalDensity.current
-
-
-    var initialPage by remember { mutableIntStateOf(0) }
-
-    val pagerState = rememberPagerState(initialPage = initialPage, pageCount = { item.size })
-
-    LaunchedEffect(key1 = Unit) {
-
-        //initialPage = Int.MAX_VALUE / 2
-
-        while (initialPage % item.size != 0) {
-            initialPage++
-        }
-        pagerState.scrollToPage(initialPage)
-    }
-
-    LaunchedEffect(key1 = pagerState.currentPage) {
-        launch {
-            while (true) {
-                delay(2000L)
-
-                withContext(NonCancellable) {
-
-                    pagerState.animateScrollToPage(pagerState.currentPage + 1)
-
-                }
-            }
-        }
-    }
-
-    Box(){
-
-        HorizontalPager(
-            state = pagerState,
-            modifier = modifier
-                .fillMaxWidth()
-                .height(400.dp)
-                .onGloballyPositioned {
-                    width = with(localDensity) {
-                        it.size.width.toDp()
-                    }
-                }
-        ) {index ->
-
-            item.getOrNull(
-                index% (item.size)
-            )?.let { item ->
-
-                Box(
-                    modifier = modifier
-                        .fillMaxHeight()
-                        .width(width)
-                        .background(Color.Black)
-                        .clickable { navController?.navigate(StimulusPostDetailRoute.route) },
-                    contentAlignment = Alignment.Center
-                ){
-
-
-                    Image(
-                        modifier = modifier
-                            .fillMaxSize()
-                            .blur(
-                                radiusX = 15.dp, radiusY = 15.dp
-                            ),
-                        painter = painterResource(id = item.coverImg),
-                        contentDescription = "",
-                        contentScale = ContentScale.Crop,
-                        alpha = 0.7f
-                    )
-
-                    Column(
-                        modifier = modifier
-                            .fillMaxHeight()
-                            .width(220.dp),
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-
-                        StimulusCoverItemPreview(item = item)
-
-                        Spacer(modifier.size(5.dp))
-
-                        Text(
-                            text = item.introText,
-                            style = TextStyle(
-                                color = Color.White,
-                                fontSize = 12.sp,
-                                fontWeight = FontWeight.Normal,
-                                textAlign = TextAlign.Center
-                            )
-                        )
-
-                        Spacer(modifier.size(5.dp))
-
-                        HorizontalDivider()
-
-                        Spacer(modifier.size(5.dp))
-
-                        Text(
-                            text = "by.${item.writer}",
-                            style = TextStyle(
-                                color = Color.White,
-                                fontSize = 12.sp,
-                                fontWeight = FontWeight.Normal,
-                                textAlign = TextAlign.Center
-                            )
-                        )
-
-                    }
-
-                }
-
-            }
-
-        }
-
-        Row(
-            Modifier
-                .align(Alignment.BottomEnd)
-                .padding(bottom = 8.dp, end = 8.dp),
-            horizontalArrangement = Arrangement.Center
-        ) {
-            repeat(pagerState.pageCount) { iteration ->
-                val color = if (pagerState.currentPage == iteration) Color.DarkGray else Color.LightGray
-                Box(
-                    modifier = Modifier
-                        .padding(2.dp)
-                        .clip(CircleShape)
-                        .background(color)
-                        .size(10.dp)
-                )
-            }
-        }
-
-    }
-
-
-
-}
 
 @Preview
 @Composable
@@ -1100,11 +977,17 @@ fun RecommendUserProfilePreview(
 
 @Preview
 @Composable
-fun RecommendUserPostListPreview(
-    modifier: Modifier = Modifier,
-    item: StimulusPostItem = StimulusPostItem(title = "이것이 제목이다", writer = "치킨 러버", coverImg = R.drawable.category3, introText = "갓생을 살고 싶어하는 당신을 위해 작성한 글!")
+fun StimulusLoadingScreen(
+    modifier: Modifier = Modifier
 ){
-
+    Box(
+        modifier = modifier
+            .background(Color.White)
+            .fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ){
+        CircularProgressIndicator(color = PurpleMain)
+    }
 }
