+++ feature/community-page/src/main/java/com/godlife/community_page/stimulus/recommended_post/RecommendedStimulusPostViewModel.kt
@@ -1,10 +1,11 @@
-package com.godlife.community_page.stimulus
+package com.godlife.community_page.stimulus.recommended_post
 
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
+import com.godlife.community_page.stimulus.StimulusPostUiState
 import com.godlife.domain.GetFamousAuthorStimulusPostUseCase
 import com.godlife.domain.GetFamousPostUseCase
 import com.godlife.domain.GetLatestStimulusPostUseCase
@@ -14,6 +15,10 @@ import com.godlife.domain.GetRecommendedStimulusPostUseCase
 import com.godlife.domain.LocalPreferenceUserUseCase
 import com.godlife.domain.ReissueUseCase
 import com.godlife.network.model.StimulusPostList
+import com.skydoves.sandwich.message
+import com.skydoves.sandwich.onError
+import com.skydoves.sandwich.onException
+import com.skydoves.sandwich.onSuccess
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -21,20 +26,10 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-sealed class StimulusPostUiState {
-    object Loading : StimulusPostUiState()
-    data class Success(val data: String) : StimulusPostUiState()
-    data class Error(val message: String) : StimulusPostUiState()
-}
-
 @HiltViewModel
-class StimulusPostViewModel @Inject constructor(
+class RecommendedStimulusPostViewModel @Inject constructor(
     private val localPreferenceUserUseCase: LocalPreferenceUserUseCase,
-    private val getLatestStimulusPostUseCase: GetLatestStimulusPostUseCase,
     private val getRecommendedStimulusPostUseCase: GetRecommendedStimulusPostUseCase,
-    private val getFamousPostUseCase: GetFamousPostUseCase,
-    private val getMostViewStimulusPostUseCase: GetMostViewStimulusPostUseCase,
-    private val getFamousAuthorStimulusPostUseCase: GetFamousAuthorStimulusPostUseCase,
     private val getPostDetailUseCase: GetPostDetailUseCase,
     private val reissueUseCase: ReissueUseCase
 ): ViewModel() {
@@ -54,12 +49,12 @@ class StimulusPostViewModel @Inject constructor(
     private val _auth = MutableStateFlow("")
     val auth: StateFlow<String> = _auth
 
-    //최신 게시물을 호출한 적이 있는지 플래그
-    private var latestFlag = mutableStateOf(false)
+    //추천 게시물
+    private val _recommendedPost = MutableStateFlow<List<StimulusPostList?>>(emptyList())
+    val recommendedPost: StateFlow<List<StimulusPostList?>> = _recommendedPost
 
-    //조회된 최신 게시물, 페이징을 이용하기에 지연 초기화
-    lateinit var latestPostList: Flow<PagingData<StimulusPostList>>
-
+    //추천 게시물을 호출했는지 플래그
+    private var isGetRecommendedPost = mutableStateOf(false)
 
     /**
      * Init
@@ -72,7 +67,8 @@ class StimulusPostViewModel @Inject constructor(
             _auth.value = "Bearer ${localPreferenceUserUseCase.getAccessToken()}"
         }
 
-        getLatestStimulusPost()
+        //추천 게시물 호출
+        getRecommendedStimulusPost()
 
     }
 
@@ -80,15 +76,28 @@ class StimulusPostViewModel @Inject constructor(
      * Functions
      */
 
-    private fun getLatestStimulusPost() {
+    private fun getRecommendedStimulusPost(){
 
-        if(!latestFlag.value) {
+        if(!isGetRecommendedPost.value){
+
             viewModelScope.launch {
-                latestPostList = getLatestStimulusPostUseCase.executeGetLatestStimulusPost()
-            }
-            latestFlag.value = true
-        }
+                val result = getRecommendedStimulusPostUseCase.executeGetRecommendedStimulusPost(auth.value)
 
+                result
+                    .onSuccess {
+                        _recommendedPost.value = data.body
+                        _uiState.value = StimulusPostUiState.Success("성공")
+                    }
+                    .onError {
+                        _uiState.value = StimulusPostUiState.Error("${this.response.code()} Error")
+                    }
+                    .onException {
+                        _uiState.value = StimulusPostUiState.Error(this.message())
+                    }
+            }
+            isGetRecommendedPost.value = true
+        }
     }
+
 
 }
